dom: use gsize type for data size parameter

+++ src/lasemrender.c
@@ -186,12 +186,12 @@ int main(int argc, char **argv)
 	if (document != NULL) {
 		if (lsm_debug_check (&lsm_debug_category_dom, LSM_DEBUG_LEVEL_LOG)) {
 			void *buffer;
-			int size;
+			gsize size;
 
 			lsm_dom_document_save_to_memory (document, &buffer, &size, NULL);
 
 			if (buffer != NULL) {
-				g_printf ("%*s\n", size, (char *) buffer);
+				g_printf ("%*s\n", (int) size, (char *) buffer);
 				g_free (buffer);
 			}
 		}
